MAC-address specific data; code clean in progress
